modified:   controller/product.go 产品功能完成 	modified:   main.go 添加测试 	modified:   models/comment.go 完善模型 	modified:   models/favorites.go 完善模型 	modified:   models/log.go 完善模型 	modified:   models/order.go 完善模型 	modified:   models/product.go 完善模型 	modified:   models/store.go 完善模型 	modified:   routers/routers.go 产品功能完成 	modified:   service/product.go 产品功能完成 	new file:   test/testData.go 添加测试 	new file:   test/testProduct.go 添加测试

+++ models/comment.go
@@ -14,9 +14,9 @@ import (
 
 type Comment struct {
 	gorm.Model
-	UserID    uint  `json:"userId"`
-	ProductID uint  `json:"productId"`
-	Likes     int64 `json:"likes"`
+	UserID    int `json:"userId"`
+	ProductID int `json:"productId"`
+	Likes     int `json:"likes"`
 }
 
 // 创建 Comment
